avoid useless conversion to set that randomizes arguments order
I admit my Python skills are pretty weak, but this set conversion
appears to be useless and it can change order of arguments from e.g.
--jobs 2 --icecream 20
to
--icecream 20 --jobs 2
As the --jobs 2 actually comes from OSC itself on multicore systems
and --icecream 20 was passed from the command line, this
obviously renders the --icecream argument rather useless.