osc add now clears files from _to_be_deleted
This happens, if a file was previously deleted, and the user changes his
mind before a commit. Such an add is no longer an add, but a cancel for the
planned delete. This resulted in unexpected loss of the file, on the server, as
the message 'is already under version control' is quite misleading in this case.