conf: detect and handle blank usernames
Fix situation where osc has added a blank username to the keyring and
update the config parse script to do the right thing and not allow blank
usernames.

I was in a very confusing situation where my keyring ended up with two
entries:

- @api.suse.de
- philipsb@api.suse.de

And I was getting 401s because of it.

Signed-off-by: Brandon Philips <bphilips@suse.de>