fix diff of files with non-ASCII/non-UTF8 chars
When diffing a file that already contains a character that
can not be decoded using ASCII or UTF-8 the diff crashes with
UnicodeDecodeError: 'utf-8' codec can't decode byte ...

The solution:

Open the source file as binary and use difflib.diff_bytes in case
we run python >= 3.5

Before we return the diff list we decode it again.

added compat_diff_bytes function to util/helper.py to support python3.0
to python3.4. In python3.3 and python3.4 the fromfile and tofile must be
concated.

compat code for python3.3 and python3.4
fix wrong dfunc call in helper.py