add support for new staging api. (command osc ps)
Create, Add, Delete staging projects
Add requests to staging

It can be issued within a project directory or with the
project as the first argument.

If you operate on a staging project or with requests the
staging project has to be given with -s and the request has
to be given with -r

You can add multiple requests / stagings.
For example:
  "-s myprj:Staging:A -s myprj:Staging:B"
  "-r 1234 -r 12345"

osc ps

For example:
    linux:~/myproject/> osc ps create
    or
    osc ps create myproject